# Implement inverse from Fermat little theorem