Mostly copied from https://github.com/Plonky3/Plonky3/pull/40, which itself is based on winterfell implementation. Sharing my results below (though they were already assessed on plonky3 related PR). I'm on  i7-9750H CPU @ 2.60GHz with Ubuntu 22.04.

![image](https://github.com/mir-protocol/plonky2/assets/30937548/770ba224-856a-4ccb-a240-ab0c774c470b)