@unzvfu 2023-08-09
Hmm, strangely I'm only seeing

> time:   [153.13 ns 153.61 ns 154.12 ns]

compared to ~140 ns in Plonky3. Are you seeing the same benchmark numbers for both?

(Also, please don't `force-push` to public branches such as PRs, it can create conflicts when someone who has checked it out tries to update.)