@Nashtare 2023-08-09
> (Also, please don't force-push to public branches such as PRs, it can create conflicts when someone who has checked it out tries to update.)

My apologies, I was fixing a typo in the description.

> Hmm, strangely I'm only seeing
> > time: [153.13 ns 153.61 ns 154.12 ns]
> 
> compared to ~140 ns in Plonky3. Are you seeing the same benchmark numbers for both?

I'm indeed getting a slower benchmark compared to the Plonky3 version. I noticed that changing the inlining from `#[inline(always)]` to `#[inline]` seemed to improve slightly improve the base case, but significantly worsen the extension fields inverse benchmarks. Is the field implementation in Plonky3 strictly identical? I'm not too familiar with it yet. 

The numbers you shared seem identical to the ones shared in https://github.com/Plonky3/Plonky3/pull/40, do you not see any changes at all? If so I'd be quite surprised, as my results above are from a comparison with the inverse currently in main, and I am definitely getting a noticeable performance improvement.