@Nashtare 2023-08-09
What's your error? And which version of nightly are you using? I'm on nightly-2023-07-01 and have no issues.