@0xKanekiKen 2023-08-09
```
Compiling plonky2_evm v0.1.1 (/home/admin/plonky2/evm)
error[E0493]: destructor of `[std::ops::RangeInclusive<u8>; N]` cannot be evaluated at compile-time
  --> evm/src/cpu/kernel/constants/exc_bitfields.rs:7:53
   |
7  | const fn u256_from_set_index_ranges<const N: usize>(ranges: [RangeInclusive<u8>; N]) -> U256 {
   |                                                     ^^^^^^ the destructor for this type cannot be evaluated in constant functions
...
29 | }
   | - value is dropped here

For more information about this error, try `rustc --explain E0493`.
error: could not compile `plonky2_evm` (lib) due to previous error
warning: build failed, waiting for other jobs to finish...
error: could not compile `plonky2_evm` (lib test) due to previous error
```

I have `1.73.0-nightly (d78bbf4bd 2023-08-03)`. Let me downgrade and try to build it again.