@0xKanekiKen 2023-08-10
This is what I was seeing yesterday:

```
Bench of your PR with #[inline(always]

Running benches/ffts.rs (target/release/deps/ffts-429ef95e0cc33801)                                              
Gnuplot not found, using plotters backend                                                                             
     Running benches/field_arithmetic.rs (target/release/deps/field_arithmetic-0703e6e5ab7fc128)                      
Gnuplot not found, using plotters backend                                                                             
try_inverse<GoldilocksField>                                                                                          
                        time:   [146.83 ns 147.16 ns 147.64 ns]                                                       
Found 7 outliers among 100 measurements (7.00%)                                                                       
  3 (3.00%) high mild                                                                                                 
  4 (4.00%) high severe                                                                                               
                                                                                                                      
Benchmarking try_inverse<QuadraticExtension<GoldilocksField>>: Collecting 100 samples in estimated 5.0010 s (8.8M iter
try_inverse<QuadraticExtension<GoldilocksField>>                                                                      
                        time:   [162.72 ns 162.95 ns 163.20 ns]                                                       
Found 7 outliers among 100 measurements (7.00%)                                                                       
  3 (3.00%) high mild                                                                                                 
  4 (4.00%) high severe                                                                                               
                                                                                                                      
Benchmarking try_inverse<QuarticExtension<GoldilocksField>>: Collecting 100 samples in estimated 5.0054 s (4.6M iterat
try_inverse<QuarticExtension<GoldilocksField>>                                                                        
                        time:   [236.98 ns 237.18 ns 237.40 ns]                                                       
Found 11 outliers among 100 measurements (11.00%)                                                                     
  6 (6.00%) high mild                                                                                                 
  5 (5.00%) high severe                                                                                               
                                                                                                                      
Benchmarking try_inverse<QuinticExtension<GoldilocksField>>: Collecting 100 samples in estimated 5.0054 s (4.0M iterat
try_inverse<QuinticExtension<GoldilocksField>>                                                                        
                        time:   [241.14 ns 241.44 ns 241.78 ns]                                                       
Found 12 outliers among 100 measurements (12.00%)                                                                     
  3 (3.00%) high mild                                                                                                 
  9 (9.00%) high severe
  ```
  
  *main*
  
  ```
  Running benches/ffts.rs (target/release/deps/ffts-429ef95e0cc33801)                                              
Gnuplot not found, using plotters backend                                                                             
     Running benches/field_arithmetic.rs (target/release/deps/field_arithmetic-0703e6e5ab7fc128)                      
Gnuplot not found, using plotters backend                                                                             
try_inverse<GoldilocksField>                                                                                          
                        time:   [157.99 ns 158.11 ns 158.27 ns]                                                       
                        change: [+7.5767% +8.0559% +8.5456%] (p = 0.00 < 0.05)                                        
                        Performance has regressed.                                                                    
Found 11 outliers among 100 measurements (11.00%)                                                                     
  5 (5.00%) high mild                                                                                                 
  6 (6.00%) high severe                                                                                               
                                                                                                                      
Benchmarking try_inverse<QuadraticExtension<GoldilocksField>>: Collecting 100 samples in estimated 5.0027 s (8.8M iter
try_inverse<QuadraticExtension<GoldilocksField>>                                                                      
                        time:   [172.34 ns 172.55 ns 172.84 ns]                                                       
                        change: [+5.4235% +6.0818% +6.7582%] (p = 0.00 < 0.05)                                        
                        Performance has regressed.                                                                    
Found 10 outliers among 100 measurements (10.00%)                                                                     
  4 (4.00%) high mild                                                                                                 
  6 (6.00%) high severe                                                                                               
                                                                                                                      
Benchmarking try_inverse<QuarticExtension<GoldilocksField>>: Collecting 100 samples in estimated 5.0005 s (4.8M iterat
try_inverse<QuarticExtension<GoldilocksField>>                                                                        
                        time:   [238.78 ns 239.06 ns 239.42 ns]                                                       
                        change: [+0.3805% +0.9980% +1.6248%] (p = 0.00 < 0.05)                                        
                        Change within noise threshold.                                                                
Found 8 outliers among 100 measurements (8.00%)                                                                       
  1 (1.00%) high mild                                                                                                 
  7 (7.00%) high severe                                                                                               
                                                                                                                      
Benchmarking try_inverse<QuinticExtension<GoldilocksField>>: Collecting 100 samples in estimated 5.0054 s (3.9M iterat
try_inverse<QuinticExtension<GoldilocksField>>                                                                        
                        time:   [249.88 ns 250.11 ns 250.38 ns]                                                       
                        change: [+3.1823% +3.8824% +4.6071%] (p = 0.00 < 0.05)                                        
                        Performance has regressed.                                                                    
Found 10 outliers among 100 measurements (10.00%)                                                                     
  3 (3.00%) high mild                                                                                                 
  7 (7.00%) high severe
  ```